requirements.txt
Flask==2.1.1
qrcode==7.3.1
Pillow==9.3.0
requests==2.26.0
Werkzeug==2.0.3
